py80211: first commit with wiphy information retrieval
This is the initial commit which currently supports retrieving
a list of wiphy objects and its information. It attempts to
provide a framework in base.py. The class nl80211_object is used
for parsing and storing attributes for a nested attribute or
an actual object in the kernel. An actual kernel object is
represented by nl80211_managed_object and has an additional
refresh() function.

The wiphy.py script uses the objects provided by base.py to
give the general idea. In the lib folder there is also a Makefile
which is used to extract information from nl80211 source files
in the kernel tree. This will update the files in the lib/generated
folder that will be added in the next commit.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>