py80211: add module to support scan functionality
New module scan.py provides a scan_request object which can be used
to prepare and send a scan request. After creating an instance the
caller can add ssids to scan for, frequencies, additional ies and
determine scan flags. The send() function will actually trigger the
scan and when the scan is finished it will obtain the scan results
and return a list of bss objects to the caller. The use of the new
objects is demonstrated in examples/example02.py. The scan requires
admin rights so need to elevate user rights.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>