py80211: base: add handling 'signed' flag in policy
The policy can only specify unsigned types. In C the signed conversion
is done in the code using a type cast. Here we have to deal with it
by adding a 'signed' flag to the policy entry affecting only the Python
side where we can deal with it doing 2s complement conversion.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>