py80211: base: allow some parameters to be None
Allow attrs and access parameters to be None. The attrs can be None so
a nl80211_managed_object can be created before retrieving information
from the kernel. Derived objects should call self.refresh() in the
constructor to obtain the attributes from the kernel. The access can
be None for lazy folks like me as it will create an instance to use
which may cause a lot of netlink sockets.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>