py80211: base: support nested attribute being list of basic types
Another use of the nested attribute type is a list of basic types. For
this the policy is checked for a 'list_type' attribute which specifies
the basic type. The attribute can be used together with the earlier
introduced 'signed' policy attribute converting all list elements to
their signed value.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>