py80211: allow custom processing of netlink attributes
The wiphy object has ATTR_TX_FRAME_TYPES (and RX) attribute which
is a two dimensional array which is wrapped in two levels of nested
attributes. Instead of coming up with a way to generically deal with
this construction we add a post_store_attrs() method to nl80211_object
which can be overriden by a derived object to do custom process. This
is used for the wiphy object to deal with the frame types attributes.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>