py80211: base: use 'map' policy attribute to mark 2-level nested list
Introducing a generic way to deal with a 2-level nested list. When
marking the policy with 'map' set True and specify 'list_type' it
will create a dictionary with nla_type() of top level nest attribute
as key which holds a list of basic typed elements as specified by
the 'list_type'. This makes the post_store_attrs() overload in wiphy
object obsolete so removing it. Still keeping the posibility for
the overload.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>